pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rawstyle"
version = "0.1.0"
description = "Apply your photographic style to Sony ARW RAW files using a reference JPEG library"
requires-python = ">=3.11"
dependencies = [
    "rawpy>=0.21.0",
    "Pillow>=10.3.0",
    "open-clip-torch>=2.24.0",
    "torch>=2.3.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "piexif>=1.1.3",
    "click>=8.1.7",
    "tqdm>=4.66.0",
]

[project.scripts]
rawstyle = "rawstyle.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rawstyle*"]
